♻️ Refactor matters settings into component
